Refactor PIC out of timer code
